ci: disable boardfarm tests on rax40
Boardfarm test on Netgear RAX40 are unstable and usually
1-2 tests fail (namely InitialApConfig and/or DevResetDefault).
These tests fail on a wide range of merge requests, so it is
likely that the root cause it outside these MRs.

Temporary disable boardfarm tests on the real device to merge
the code. This commit needs to be reverted after the root cause
is found.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>