bwl: dwpal: read up to five external events at a time
In the current implementation, it is seen that if a burst of client
association events is received from the hostapd, the ap manager is not
quick enough to read the clients information and the hostapd buffer is
full which may cause the hostapd to be stuck until it is read since the
hostapd writes are blocking.

When the hostapd is in that state it is not responding to association
requests and some clients may fail to associate.

By increasing the number of events to be handled in each select, this
scenario is less prone to happen. The limit of five events was
arbitrarily selected and was found sufficient for a burst of 200 clients
on one hand and didn't cause any other functional issues on the other
hand.

Signed-off-by: Adam Dov <adamx.dov@intel.com>