agent: add ability to abort in-progress ZW-DFS flows
preparative commit.

This commit is a first in a series of commits which
adds the ability to cancel an in-progress ZW-DFS flow in cases where
an external channel switch was detected (triggered by using hostapd_cli
for example).

If an external channel switch is detected, the current ZW-DFS flow will
be aborted, releasing the ZW-DFS antenna and give way to the external
channel switch being requested.

- add abort_zwdfs_flow() function to channel selection task.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>