agent: abort in-progress ZW-DFS flow on unexpected CAC_Started event
The "CAC_Started" event is raised by the agent's primary radio when a
channel switch is performed into a usable DFS channel.

If this event is received unexpectedly, it serves as an indication for
an external channel switch, in which case, the current ZW-DFS flow will
be aborted, releasing the ZW-DFS antenna and give way to the external
channel switch being requested.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>