framework: bpl: refactor cfg_get_all_prplmesh_wifi_interfaces
The existing implementation of cfg_get_all_prplmesh_wifi_interfaces()
couples the section names (radioX) to the index used. This means that if
we have 3 radio their section names in the config file must be radio0,
radio1,radio2.
In practice this limitation is incorrect.
Moreover, we may use the same configuration file to various platform
modes wherein one mode the first 3 radio-sections are relevant and in
another some other sections are relevant.

- Decouple the section name from the way we read the list of configured
hostapd interfaces.
- The API returns all listed hostap interfaces and the calling
implementation filters the not useable/available interfaces.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>