CMakeLists: move cache parameters defaults to end
In CMake the cache parameters will not be set if the parameter already
exists/is-set. Setting these parameters to their default values before
the per-platform settings mean that the per-platform settings will not
take effect.

Moved the default configuration to after the per-platform settings to
set only parameters that are not set for platform-specific settings.

Note that this change also aligns with the required support of not
overriding the settings from the shell.

Signed-off-by: Adam Dov <adamx.dov@intel.com>