tasks: interface, messages and Task
This commit introduces the following -

changes:
* task pool interface - enum of messages that are used
in the following commits.
* task - interface to handle-event of enum and shared-ptr

addition:
task messages - a common place to messages structures that
are sent between tasks.
The reason these messages have a dedicated place is because
they are not belong to a specific task, rather, these
are the messages that are sent _between_ tasks, they
belong to more than one task, they have their own place.
e.g TaskA ---<messageX>--> TaskB. messageX is known in
both TaskA and taskB.

Signed-off-by: Ran Regev <ran.regev@devalore.com>