transport: add SHMGET_MID and transport support
preparative commit

the communal ack between indepednant process requires adapatation
to the transport, so the old MID will be semi replaced with the
shm  mid.

if shm mid < 65536 it'll be treated as normal mid
else it'll be registered with the pair of key and value
key is dest + normal mid, value is the shm mid, so on response on he
transport this will be converted to the mid and sent back to the
controller

the problem here is that sem_wait function that supposed
to sync read and write calls to this function, is blocking
for that we'd hope to have a thread of some sort to not hang
the transport.

- adds functions:

get_next_message_id
SHMEM_set_next_message_id
get_rid

getRidCounter
incRidCounter

Signed-off-by: Daniel Edri <daniel.edri@devalore.com>