backhaul_manager: fix CMDU forwarding to zwdfs slave
As part of extended refactoring that replaced the use of sockets to the
use for file-descriptors, a bug was introduced which prevents forwarding
of VS CMDUs to the ZWDFS slave. This is since it is not listed in the
slaves_sockets, but in the m_disabled_slave_sockets.

Added check in the m_disabled_slave_sockets when the fd is not found in
the slaves_sockets.

Signed-off-by: Adam Dov <adamx.dov@intel.com>