tests: add a delay between CAPI commands
The current UCC listener implementation seems to have an issue when
commands are sent in a very rapid succession: it makes the boardfarm
tests hang.

Instead of changing the multiple tests that do this, add a delay in
cmd_reply() itself.

This is a workaround to try to mitigate the actual issue until it can
be properly fixed.

See https://jira.prplfoundation.org/browse/PPM-925 for the full
details about the issue.

While at it, reduce the timeout for the boardfarm tests to 30
minutes (instead of 4 hours from the project settings). Since the
boardfarm tests usually take around 4 minutes to complete, this will
at least reduce the effects a hanging job has on the overall CI.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>