docker: openwrt: add a more advanced cache mechanism
Similar like the RDK-B build, add a -c option to build.sh to specify a
cache directory. After the build, the contents of the dl directory are
copied into the cache directory. Before the build, the contents of the
cache directory are copied into the docker container.

To avoid invalidating the cache, the timestamp of the cache tarball is
set to a fixed value (docker only uses the mtime of the file to
determine if it needs to be refreshed or not). If we don't do this,
running the same build twice in a row will invalidate the docker cache
because the cache tarball has changed.

The cache is per-target, but the first target that is built seeds a
"generic" cache that is used for other targets as well. Most downloads
are anyway the same between different targets.

For now, only downloads are cached. This could be extended later with
ccache.

Note that for now, there is nothing that limits the size of the cache.
For the downloads this is not a limitation in practice, since most
packages are not updated very often. When we add ccache, it's also not a
problem because ccache manages its own size.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>