agent db: dfs state
Since there is a need to know for a channel
the duration the channel is in its current dfs state:
this commit hides the public dfs_state member under get() and set()
functions. The set() function takes care for saving the time the dfs state
was changed, by simply marking it as now() each time the state is
changed via the set() function - which this commit made it
the only way to set the dfs state.

Changed all places in the code that used dfs_state directly.

Signed-off-by: Ran Regev <ran.regev@devalore.com>