prplmesh_utils: main agent status: check only for OPERATIONAL
Main agent test uses logs to check for transition from
CONNECTED to OPERATIONAL state.
Recently a new state has been added between these two
and it caused boardfarm tests to fail.

Check only for the transition to OPERATIONAL state
regardless of the previous state. It will help to
avoid issues like this in the future.

Signed-off-by: Vitalii Komisarenko <vitalii.komisarenko@mind.be>