tests: boardfarm: changed vap>sta logic to sta>vap on wifi connections
Previously when an association between a device and a VAP was made,
a associate/disassociate function on the VAP object would be called
to make the connection. This commit changes the logic so the device
is the one being called on the process, making it easier to track
the VAP association when connecting to wifi.

Signed-off-by: David <davidmgafonso@gmail.com>