nl80211: implement set_channel
set_channel is used to configure hostapd with the correct channel when
the wireless backhaul link is established. Without it, hostapd keeps
its previous configuration, and might not be on the channel that the
supplicant configures. This can cause issues where hostapd will tries
to use channel 48 for example, while the radio is on channel 36 and
refuses to switch to 48 because the wireless backhaul link is
established.

In theory this should mean that in wireless backhaul tests the AP is
never beaconing while the supplicant is connected on the same
radio. For some reason, in practice we can see the AP is beaconing
most of the time, and that only from time to time the test fails
because the AP is not beaconing. Still, we implement set_channel in
this commit to see if the situation improves.

Implement set_channel by taking the dwpal implementation (since the
hostapd configuration is the same in that regard). However, since we
modify the hostapd configuration file in the nl80211 implementation,
do it here as well. This means instead of using 'SET' and 'UPDATE', we
load the configuration, change the relevant parameters, store it back
and issue 'UPDATE'.

While at it, remove the handling of 80+80 in the dwpal implementation,
because the current driver doesn't support it.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>