boardfarm: check_cmdu_type_single: add more debugging info
When debugging test failures like in [1], it can help to have more
information about what went wrong.

- When multiple CMDUs are found, print them. This should allow us to
know more easily which packets it detected as duplicate.

- When checking for a CMDUs, also print the destination MAC, and the
message ID. This makes it easier to understand what is being
checked. By default the destination MAC is the multicast address, so
it's print as "Multicast". If there is no message id, it's printed as
"Any".

[1]: https://jira.prplfoundation.org/browse/PPM-962

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>