controller: fix combined infra metrics MAC addresses
In PPM-213, MAP-5.7.1 is failing because the Combined Infrastructure
Metrics message does not contain the correct link metrics data.

There were two issues:

- The responder and neighbor MAC addresses were always the same. This
turned out to be due to an error where the link metrics are
stored. Instead of storing '[responder][responder]', store
'[responder][neighbor]'. Also check that the Tx and Rx neighbor MAC
addresses are the same, like it's done for the responder.

- The receiving and neighbor MAC addresses pair that follow were
always all-zeroes. This was due to the 'interface_tuple' variable
being copied by value, which in turns means we were only storing the
MAC addresses in a temporary object that was destroyed immediately
after. Since a temporary variable is anyway not needed, fix it by
accessing the object directly.

This should fix PPM-213, although there might remain some stability
issues with the certified devices.

This was actually caught by cppcheck (redundantInitialization). Update
the cppcheck issues since it's now solved.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>