ci: add a release mechanism
It is based Gitlab's example of pusblishing assets for releases ([1]).

Note that while it's also possible to use the "release" keyword in the
Gitlab configuration file ([2]) in the end it translates into a shell
script that calls the same release-cli tool. By calling it ourselves,
we have more control over what it does, and more importantly it allows
us to attach assets (e.g. binaries) to the release.

The builds for the Axepoint, Netgear Rax40, NEC WX3000HP, Turris
Omnia, and GL.iNet B1300 are uploaded. Both the prplmesh ipks, and the
full images are included.

Because the job needs access to the artifacts for the prplwrt devices,
it's easier to have the release job at the end of 'platform'
subpipeline.

The actual release is a two-step process: upload the artifacts that we
want to attach to the release, then create the release. Uploading the
artifacts is done with Gitlab's "generic packages" (see [3]). Then,
release-cli is used to create the release and attach links to the
packages.

The 'release' job has a 'needs' on the 'upload-release-artifacts'
job. Since they are in a different stage anyway, this is not necessary
for the release job to wait until the upload is finished. It does
however prevent the release job from being skipped if a test
failed (for example, if boardfarm tests on the rax40 failed).

Note that the URL used for "generic packages" has a very specific
format:
/projects/:id/packages/generic/:package_name/:package_version/:file_name

Specifically, 'package_name' only accepts specific characters. See [4]
for details. Similarly, 'package_version' has to match the following
regex: '/\A\d+\.\d+\.\d+\z/' (i.e. it should only contain 3 numbers
separated by dots). Since we use the git tag as-is for the package
version, this means they also have to match this regex.

For the Gitlab releases documentation, see [5].

[1]: https://gitlab.com/gitlab-org/release-cli/-/tree/master/docs/examples/release-assets-as-generic-package/
[2]: https://docs.gitlab.com/ee/ci/yaml/#release
[3]: https://docs.gitlab.com/ee/user/packages/generic_packages/
[4]: https://docs.gitlab.com/ee/user/packages/generic_packages/#publish-a-package-file
[5]: https://docs.gitlab.com/ee/user/project/releases/index.html

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>