controller:nbapi: fix appear of parameters of security object
The 'filtering' logic in the %populate section of data model
did not work with current version of Ambiorix as it's suppose to
work.

Change logic of filtering mechanism in the ambiorix event
implementations. Now, it's invoke event every time when any object
is changed. And when 'Controller.Network.AccessPoint.*.Security.ModeEnabled'
set to WPA2-Personal it adds or removes parameters () accordingly.

PPM-831
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>