Revert "docker: openwrt: add a more advanced cache mechanism"
This reverts commit 74c3c565d5a462e0e57505b82de680d6d8a03ada.

On recent versions of Docker, the ADD command no longer uses a timestamp
to determine if it's old or not; instead it uses a hash of the file.

There seems to be no way to disable this behaviour.

With no better solution in sight, simply revert the commit that adds the
cache.

Fixes: https://jira.prplfoundation.org/browse/PPM-977

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>