ucc: Fix message type validation error print
On UCC dev_send_1905 command validation we check that the requested
message type is valid. When validation fails we print an error message
with the recieved string with addition of `0x` prefix. This addition
causes the print look as if the format is was really wrong i.e:
"0x0x8020".
So remove the addition of 0x to the string.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>