son_slave: handle Channel Scan Reporting Policy
PPM-359
R2 test adds a new TLV to the MULTI_AP_POLICY_CONFIG_REQUEST_MESSAGE
to set the Channel Scan Reporting Policy.
The requirement was to set the value to the AgentDB for later use.
This commit adds reception of `tlvChannelScanReportingPolicy` in son_slave
and setting a new value `report_indepent_scans_policy` to the radio in
the AgentDB.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>