tlvf: support python2 build for ugw platform
The commit 933a634 added explicit use of python3, but since when the
prplmesh is built by the UGW platform (openWrt based) it uses python2
for the build, it breaks the build for that platform.

Reverted the change only for the UGW platform.

For other platform it seems the exceute_process isn't working properly
so changed to the "set" command instead.

Signed-off-by: Adam Dov <adamx.dov@intel.com>