transport: reliable multicast: do not send to the local agent
In PPM-973, we noticed some packets are sent with the same source and
destination MAC addresses.

By looking at the transport logs with DEBUG level, it turns out that
the local agent AL-MAC is in fact in the neighbors map, which causes
us to send packets with same source and destination.

When this happens in docker tests, the kernel logs contains the
following line multiple times:
kernel: br-lan: received packet on eth1 with own address as source address (addr:66:a1:e1:aa:51:1e, vlan:0)

If source and destination MACs are the same, just continue to the next
neighbor.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>