boardfarm: bwl: do not fail on timeout
Commit 301b54be8b9eae016e87b4f505961a0c32e4f92d introduced a timeout
for the bwl events to avoid them from blocking in case there is no
listener.

In practice, it turns out it sometimes fails with exit code
124 (i.e. 'timeout' reached its timeout) which makes the test
fail (for example: [1]). When looking at the log files, it turns out
the slave received the event and processed it correctly, so there is
no reason the test should fail in this case.

Add '|| true' to the end of the command to make sure it never
fails.

[1]: https://gitlab.com/prpl-foundation/prplmesh/prplMesh/-/jobs/960997542

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>