ucc: Add support for odd number of digits
CAPI commands HEX notation is usualy sent with even number of digits as
we expect that every octet consist of 2 digits.
Practically, we already had issues where the number of digits were odd
for example `0x0` value.
When inpecting the R2 test spec we did find examples of those odd values
(for example section 4.14.1 step 3).
so this commits add the support for those values.

Signed-off-by: Semyon Barenboym <semyon.barenboym@devalore.com>