openwrt: axepoint/rax40: update to include the ugw-8.4.2 changes
OPENWRT_VERSION is updated to use the latest hash of
feature/PPM-339-ugw-8.4.2-release-prplmesh. This branch is a rebase of
our changes on top of current prplwrt/prplwrt.

For the intel feeds, we just track upstream. We still need to keep it in
the netgear-rax40 profile because our jq script overrides the feeds list
entirely.

feed-prpl is also upstream.

The syntax of the profile yaml files has changed a litte: hash ->
revision and add the tracking_branch. Note that tracking_branch is just
documentation, it's not used by the script.

The source directory of bridge-utils changed. However, it is now also
properly installed to the staging directory, so we can use normal
include/library. Apply a patch by Petr that does exactly this. Note that
this will cause failures on older UGW.

The sFourAddrMode hostapd parameter was removed from UGW 8.4.2, so don't
set it any more for backhaul VAPs.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>
Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Alex Kanter <alex.kanter@devalore.com>
Signed-off-by: Petr Štetiar <ynezz@true.cz>