tests: boardfarm: added get_vap function for both docker and real device
Added a bwl output for the vap information on update_vap_credentials
to avoid getting the information from the log files
The information given is based on the available_vaps id so the postions
are kept when writing the file. Some of test would break otherwise
because they rely on a specific vap order. This should be refactored
in the future

Signed-off-by: David <davidmgafonso@gmail.com>