wireless utils: find all overlapping channels
Added a function to find opverlapping channels.
Find pairs of overlapping channels and bandwidths.

This functionality is required for example
when we report about operating-class and channels that
are in a specific dfs state. For the given channel-bandwidth we
first need to find all the opverlapping channels/bandwidths
(the introduced functionality in this commit) and then we can
compose operating-class/channel pairs from them.

Signed-off-by: Ran Regev <ran.regev@devalore.com>