ci: move individual controller tests to axepoint
Currently, the dummy controller tests cannot be run on the PTK
testbed (we are able to run them on the Essensium testbed only because
the ethernet switch used is different).

Since we anyway can run them on a device, move them to the
axepoint. The axepoint was chosen because it's the only device present
in both PTK and Essensium certification testbeds.

Add a separate stage for them to have a separate column in the Gitlab
UI.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>