docker: rdk-b: add support for turris-omnia
Since the Turris Omnia support is still on a side-branch in RDK-B, with
a specific manifest file that doesn't contain Raspberry Pi, replace
raspberrypi with turris - including in the CI job.

Instructions for flashing the Turris Omnia are a bit sparse [1] but
apparently we need the wic file.

Since there is already a build directory called turris-omnia that
contains the prplwrt build, append -rdk to the build directory name.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>

[1] https://wiki.rdkcentral.com/pages/viewpage.action?pageId=114986683