agent: handle DCS propagation to Backhaul.
Preparative commit.

Currently only the controller sends Channel-Scan requests and has
the agent forward notifications to it.
With the new DCS R2 channel scan task in the agent, the son_slave needs
to receive requests from the backhaul & propagate the notifications to
it as well.
The incoming CMDUs in the backhaul are forwarded to the Channel Scan
Task automatically by the backhaul manager and will be handled within.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>