controller: reading scan params from DB on scan request build
The scan parameters: "dwell_time_msec" and "channel_scan_pool"
should be read from the DB when building the scan request message
instead of when a scan request event is received.

The reading of this parameters was moved forward in the flow in order
to read the most up-to-date parameters from the DB, in case  of
parameters change between the triggering of the scan request and the
sending of the scan request message by the task.

- Moved the reading of dwell_time_msec from
"handle_scan_request_event()" to "trigger_pending_scan_requests()"

- Moved the reading of curr_channel_pool from
"handle_scan_request_event()" to "add_operating_classes_to_radio()"

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>