ci: apply UCI settings for scheduled certification tests
'1f3028d502eaf8897fc6bbd812b6f0385277a00d' added a script to apply UCI
settings to devices, but it was mistakenly not called for scheduled
certification tests. Add it now.

While at it, move it to after deploy_firmware since it makes more
sense to apply those settings after the image is deployed. In theory
it could even be moved after deploy_ipk, but that's not possible until
PPM-312 is fixed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>