nbapi:on_action: fix the way of sending ap_config_renew_msg
The previous way of sending ap_config_renew message did not work properly.
The values of network key are removed when it's not need it.

Modify the way of sending  ap_config_renew message.
Comment out line which removes values from network keys.

Also, skip adding BSS to the database if authentication is not open
and the network key isn't present.

PPM-1009
--
Signed-off-by: Krystyna Oliinyk <k.oliinyk@inango-systems.com>