channel selection task: added channel-list-ready
Sending channel list ready event from
channel-selection-task when the channel list
is ready.

This is another place (like cac-started, csa and more)
that we just freely sending a task-event to be catched by whoever
needs it. In this case we signal other tasks that the channel list
is ready. The first customer of this message is the cac-task that
must have the list ready before it can do anything.

Signed-off-by: Ran Regev <ran.regev@devalore.com>