bwl: Configure profile1/2 disallow on hostapd configuration file
Add the profile1/2 disallow flags to "update_vap_credentials()" and
"set_vap_multiap_mode()" functions which edit the bss configuration on
the hostapd configruation file.
Since we need to know the profile disallow flag which is important for
knowing how to configure the VLANs, preset the VAPs properties in
the m_radio_info.available_vaps.

Since we currently can't configure the profile-1/2 disallow flag on the
nl80211 flavor of BWL, port these changes only to BWL DWPAL.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>