bwl: Override BSS parameters only if they don't exist
On the previous commit, the VAP configuration was preset to the
"m_radio_info.available_vaps" container.
Since we don't want to override it with partial data, override it only if the
detected VAP does not have preset information.

In addition, removed the use of "beerocks::string_utils::bool_str" which is
not needed anymore, and simplify the nl80211 implementation by using
lambda function.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>