fronthaul: Add the Profile-1/2 to copy_vaps_info
The function copy_vaps_info() copy the BSSIDs parameter to the sVapInfo which is
used to send the BSSIDs information to the son_slaves.
Copy to that struct the newly added profile-1/2 disallow parameter and the bBSS flag
that was missing.

Also, remove the use of string_utils::bool_str() from a print, since it is no longer
required for printing a boolean.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>