ci: merge the list of passing tests
Having separate files for controller, agent, and backhaul type can be
convenient when we only want to test some of them in the scheduled
tests.

However, in practice we run all of them everyday, and it adds a lot of
complexity in the Gitlab schedules (we have 3 times more pipelines per
testbed, per branch we want to test).

Merge all the passing tests in a single file.

Split the list of passing tests into two variants. Variant A consist
of the tests with even line numbers in the existing list of passing
tests, while variant B consist of tests with odd line numbers. Only
one of them will be run each day.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>