ci: merge the list of passing tests
Having separate files for controller, agent, and backhaul type can be
convenient when we only want to test some of them in the scheduled
tests.

However, in practice we run all of them everyday, and it adds a lot of
complexity in the Gitlab schedules (we have 3 times more pipelines per
testbed, per branch we want to test).

Merge all the passing tests in a single file.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>