ci: split passing tests into variants
The current nightly tests take too long to complete: they consist of
114 tests, and take about 19 hours in total. As a consequence, the
certification testbed is rarely available during the day to test merge
requests.

Split the list of passing tests into two variants. Variant A consist
of the tests with even line numbers in the existing list of passing
tests, while variant B consist of tests with odd line numbers. Only
one of them will be run each day.

This allows us to test almost everything in a single day. For example,
for 4.4.2 the variant A has only 5GL, but for 4.4.3 it has both 24G
and 5GH. Since most of the tests behave the same way in the different
frequencies and we still test the two radios (in different tests),
this means we keep an acceptable coverage.

There are exceptions to this: some controller tests (e.g. MAP-5.3.1),
and one agent test are unique (i.e. they don't have different
frequencies). For those, they are added manually to each file so that
we still run then every day.

4.7.9 is another exception: for now it's only in the variant
A. However, this is because its wireless
counterpart (MAP-4.7.9_BHWIFI) is not passing yet. With the work on
PPM-982 it will be added relatively soon to the variant B file so we
don't have to add it manually.

The following snippets can be used to re-generate each file.

Variant A:

{ awk 'NR%2==0' passing_tests.txt &&
  printf 'MAP-4.2.1\nMAP-5.3.1\nMAP-5.5.1\nMAP-5.6.1\nMAP-5.7.1\nMAP-5.10.1\nMAP-5.10.2\n';
} | sort -u > passing_tests_variant_A.txt

Variant B:

{ awk 'NR%2==1' passing_tests.txt &&
  printf 'MAP-4.2.1\nMAP-5.3.1\nMAP-5.5.1\nMAP-5.6.1\nMAP-5.7.1\nMAP-5.10.1\nMAP-5.10.2\n';
} | sort -u > passing_tests_variant_B.txt

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>