controller: nbapi: get_param_string: fix SEGV
When a parameter doesn't have a default value, it's default value (as a
string) will be a NULL pointer. Standard C++ doesn't accept a NULL
pointer as the initialiser of a string, which gives a segmentation fault
when returning param_val.

Also, amxc_var_dyncast directly returns a pointer to the internal
variable, and amxc_var_clean frees that variable. Thus, we were creating
the string from freed memory. This usually works because nothing else is
allocated in-between, but sometimes fails.

Fix this by using a std::string as the local variable, and assigning to
it only if the pointer is non-NULL.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>