agent: channel scan task: is scan request finished
Iterate over the found Radio Scans within the given request.
If a Radio Scan is not yet in the SCAN_DONE or SCAN_FAILED states, it
means that their scans are in progress or have yet to be triggered.
Either way this means that the whole request is still not finished.

- Add "is scan request finished" implementation.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>