agent: channel scan task: trigger next radio scan
Iterate over the found Radio Scans within the given request.
Find the first Radio Scan that is in the PENDING_TRIGGER state and
trigger a channel scan for it.
Currently only one Channel Scan per radio is supported.
When PPM-711 is resolved we need to rework the trigger mechanism to
support multiple simultaneous radio scans.

- Add "trigger next radio scan" implementation.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>