agent: channel scan task: trigger radio scan
For a given Radio Scan, create and send a TRIGGER_CHANNEL_SCAN CMDU
message to the monitor.
Currently the TRIGGER_CHANNEL_SCAN message requires a Dwell-Time
parameter, but after PPM-1040 is resolved we will be able to send a
request without a set Dwell-Time, and instead use the pre-configured
Dwell-time in the driver.

- Add "trigger radio scan" implementation.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>