agent: channel scan task: store results in db
When fresh scan results are received from the monitor the task is
responsible for storing and maintaining the results in the agent's DB.
Old results need to be cleared before new results are inserted.

It is important to note that not all the results are cleared, only the
results of channels which receive new results are cleared.

    If a scan was requested on channels 36 & 52,
    and there are stored results for channels 36, 40 & 52,
    and from the scan we received results for only channel 36.
    The stored results for channel 36 will be cleared, while the stored
    results for channels 40 & 52 will remain intact.

- Check if the currently stored results need to be cleared or not.
- Add "store results in db" implementation.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>