agent: channel scan task: send report to controller
According to Multi-AP Specification Version 2.0, section 10.2.2.2,
a CHANNEL_SCAN_REPORT_MESSAGE needs to be sent when the associated
request's Radio Scans are done.
The CHANNEL_SCAN_REPORT_MESSAGE is described in Multi-AP Specification Version 2.0
section 17.1.34, and consists of one Timestamp TLV (17.2.41) and one or
more Channel Scan Results TLVs (17.2.40).

- Create a CHANNEL_SCAN_REPORT_MESSAGE CMDU.
- Add a Timestamp TLV to the CMDU.
- Iterate over the given reqeust's radios
-- If a fresh scan was requested, fill the Channel Scan Results TLVs
from the Agent's DB with the request's corresponding results.
-- If a fresh scan was not requested, fill the Channel Scan Results TLVs
with all the stored results for the selected radios.
- Send CHANNEL_SCAN_REPORT_MESSAGE

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>