agent: channel scan task: finalize task work flow
The Channel Scan task need to handle a pending request queue, a running
Radio Scan, and an active request at the same time.

When there is a running Radio Scan, it would manage an
internal state, and go though a StateMachine.

After handling the current scan's state, the work() function checks if
the currently active request is ready to send a report back to it's
requester.

If there is no currently running scan, the work() function needs to
check if there are any pending requests, if there are, the next scan
will be triggered.

PPM-363.

Signed-off-by: Itay Elenzweig <itayx.elenzweig@intel.com>