ci: split the nightly certification tests variables
Until now we always created two different pipelines to run nightly
certification tests on the ISR and Essensium testbed. This not only
resulted in unnecessary builds being created, it also made the Gitlab
schedules harder to maintain (since for every test suite we needed two
scheduled pipelines).

The following changes are done:

- run-certification-tests-platform: append the testbed name to the end
of the $TESTS_TO_RUN and $DEVICE_UNDER_TEST variables. This will allow
us to specify different tests to be executed on the different testbeds
in a single pipeline. Because it's not possible to use variables
inside other variables (see [1]) we create a file that will later be
used to populate the right variables.

Because the TESTS_TO_RUN_X variables will be set from the parent
pipeline, they will always be set in the child pipeline (if they were
unset in the parent they will be set to a dollar sign followed by the
variable name). To work around it, check that the variable does not
begin with a dollar sign.

- run-certification-tests-prplmesh: do the same changes as for
'run-certification-tests-platform'. Note that the regex for
TESTS_TO_RUN_X is not strictly required, but to make sure the job
doesn't accidentally start if it was ever moved to the child pipeline,
add it here as well.

- platform: propagate the new TESTS_TO_RUN_{ESSENSIUM,ISR} and
DEVICE_UNDER_TEST{ESSENSIUM,ISR} to the child pipeline.

- .run-certification-tests: drop $TESTS_TO_RUN and $DEVICE_UNDER_TEST,
and instead read the two files to get the list of tests and the DUT.

[1]: https://gitlab.com/gitlab-org/gitlab-runner/-/issues/1809

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>