controller: trigger pending scan requests - abort when fail
Change trigger_pending_scan_requests() function recovery behavior when
the creation of CMDU messages or internal message TLV fail.

Currently, upon failure to create the CMDU / TLV - the mid of
radio_scan_request will be sed as invalid and the agent status will be
set as idle.

This will cause the tasks to re-try sending the message to the agent,
and possibly fail again to create the CMDU / TLV.

This commit change recovery behavior. The function will now abort the
requests of a given agent and report the channel status as fail.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>