controller: db: factor out dm_set_radio_bss()
As part of the add_vap function, the BSS datamodel is updated. Since
we're going to add another code path where this has to be done, refactor
this into a new function, dm_set_radio_bss.

The code has moved to be close to the other dm_ functions. Otherwise,
the only change is converting the sMacAddr arguments to strings.

Also remove spurious EOL whitespace (automatic by editor).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>