controller: handle ApOperationalBSS in topology response
The Topology Response message contains an AP Operational BSS TLV that
updates the information about the BSSes. Use this information to update
the VAP info in the database.

To support non-prplMesh agents, a VAP ID must be generated if it doesn't
exist yet. We just take the maximum of the existing VAPs, plus one.

https://jira.prplfoundation.org/browse/PPM-918

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>