agent: send topology notification when VAPs list is updated
When the VAPs list is updated, this changes the information in the
topology response message. In such a situation, we must send out a
toplogy notification (with no additional TLVs). Do that.

Note that we always send a topology notification, even if the
information in the VAPs list hasn't actually changed. This may cause a
spurious topology query/response cycle, but the overhead of that is
acceptable.

https://jira.prplfoundation.org/browse/PPM-918

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>