controller: skip checking of mid and ACK as response to scan request
Currently, the MID (message id) of outgoing messages is unknown at the
time of sending and therefore cannot be checked against when receiving
ACK messages or scan_report events.

As a result the controller is unable to pass the ACK message to the
corresponding controller task.

This commit remove the use of ACK messages in
dynamic_channel_selection_r2_task as validation for scan request
reception by the agent. The scan request message is assumed to be
received without ACK.

This commit also remove the checking of mid_to_agent_map when receiving
scan reports. If the radio status is SCAN_IN_PROGRESS,
it is assumed to expect the scan report of this agent.

Signed-off-by: Oren Vormaser <oren.vormaser@devalore.com>