common: prplmesh_utils.sh: abort if ebtables failed
In PPM-973, some boardfarm tests would sometimes fail due to messages
being lost.

It turns out that the actual issue was that the ebtables rules could
not be applied, because the kernel module was not loaded.

Because of this, some of the relayed multicast messages were being
sent back to the sender in a loop, which would flood the transport for
a short amount of time during which other messages were discarded.

This situation has been going on for months, showing that the problems
introduced by the missing ebtables rule can be subtle and easily
overlooked.

For the particular case of the boardfarm tests in docker, this has
been solved by loading the ebtables module at boot time on the host.

To make sure this doesn't go unnoticed again in the future, abort
starting prplMesh if the ebtables was not added correctly.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>